corrected some duration formatting issue and added a toolbar to the settings screen

+++ app/src/main/java/com/punisher/fitnesstracker/dto/FitnessActivity.java
@@ -80,7 +80,7 @@ public class FitnessActivity {
      * calculate the average time for this FitnessActivity
      * @return the average time
      */
-    public float getAverage() {
+    public int getAverage() {
         float speed = (float)_distance / (float)_duration;
         float result = 1000 / speed;
         return Math.round(result);
